+++ lib/lutaml/model/schema/definitions/simple_content.rb
@@ -9,12 +9,11 @@ module Lutaml
         # `attribute :content, :<base_type>` and any additional
         # attributes after the regular members. RNG leaves this nil.
         class SimpleContent
-          attr_accessor :base_class, :additional_attributes, :required_files
+          attr_accessor :base_class, :additional_attributes
 
-          def initialize(base_class:, additional_attributes: [], required_files: [])
+          def initialize(base_class:, additional_attributes: [])
             @base_class = base_class
             @additional_attributes = additional_attributes
-            @required_files = required_files
           end
         end
       end
