+++ lib/lutaml/model/schema/renderers.rb
@@ -4,7 +4,8 @@ module Lutaml
   module Model
     module Schema
       module Renderers
-        autoload :Registration,   "#{__dir__}/renderers/registration"
+        autoload :Registration, "#{__dir__}/renderers/registration"
+        autoload :RequiredFilesCalculator, "#{__dir__}/renderers/required_files_calculator"
         autoload :MemberDecls,    "#{__dir__}/renderers/member_decls"
         autoload :Mappings,       "#{__dir__}/renderers/mappings"
         autoload :Model,          "#{__dir__}/renderers/model"
